demo/cli/requirements.txt
environs==9.5.0
Flask==2.2.2
grpcio==1.48.1
grpcio-health-checking==1.48.1
grpcio-reflection==1.48.1
opentelemetry-api==1.14.0
opentelemetry-sdk==1.14.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-exporter-zipkin-proto-http==1.14.0
opentelemetry-instrumentation-grpc==0.35b0
profanity-filter==1.3.3
protobuf==3.19.4
python-logging-loki==0.3.1
spacy==2.3.9

accelbyte-py-sdk
bitarray
httpx
mmh3
PyJWT[crypto]
PyYAML
requests
websockets
python-dotenv
